refactor(visual-review): consolidate shared GitHub integration code
Move GitHubRateLimitError and GITHUB_API_VERSION from visual_review's
local github.py into posthog/models/integration.py alongside
GitHubIntegrationError so both visual_review and tasks share a single
source of truth.

Add get_access_token() to GitHubIntegration to encapsulate the
refresh-then-read pattern, replacing seven direct
sensitive_config["access_token"] accesses and an explicit
access_token_expired()/refresh_access_token() pair in logic.py.

Replace ten hardcoded "2022-11-28" strings in integration.py with the
new GITHUB_API_VERSION constant.
