+++ products/visual_review/backend/logic.py
@@ -23,10 +23,11 @@ import structlog
 if TYPE_CHECKING:
     from posthog.models.integration import GitHubIntegration
 
+from posthog.models.integration import GitHubRateLimitError  # noqa: F401 — re-exported for facade
+
 from .classifier import SnapshotClassifier
 from .db import WRITER_DB
 from .facade.enums import ReviewDecision, ReviewState, RunPurpose, RunStatus, SnapshotResult, ToleratedReason
-from .github import GitHubRateLimitError  # noqa: F401 — re-exported for facade
 from .models import Artifact, QuarantinedIdentifier, Repo, Run, RunSnapshot, ToleratedHash
 from .signing import sign_snapshot_hash, verify_signed_hash
 from .storage import ArtifactStorage
@@ -360,7 +361,7 @@ def _get_merge_base_sha(github: GitHubIntegration, repo_full_name: str, base: st
 
     from .github import github_request
 
-    access_token = github.integration.sensitive_config["access_token"]
+    access_token = github.get_access_token()
     try:
         response = github_request(
             "GET",
@@ -399,7 +400,7 @@ def _get_default_branch(github: GitHubIntegration, repo_full_name: str) -> str:
 
     from .github import github_request
 
-    access_token = github.integration.sensitive_config["access_token"]
+    access_token = github.get_access_token()
     try:
         response = github_request(
             "GET",
@@ -1148,7 +1149,7 @@ def _resolve_repo_by_id(github, repo_external_id: int) -> str | None:
     """
     from .github import github_request
 
-    access_token = github.integration.sensitive_config["access_token"]
+    access_token = github.get_access_token()
     response = github_request(
         "GET",
         f"https://api.github.com/repositories/{repo_external_id}",
@@ -1181,10 +1182,7 @@ def _github_api_request(
     safe_path = "/".join(quote(segment, safe="") for segment in path.split("/"))
 
     github = get_github_integration_for_repo(repo)
-    if github.access_token_expired():
-        github.refresh_access_token()
-
-    access_token = github.integration.sensitive_config["access_token"]
+    access_token = github.get_access_token()
 
     url = f"https://api.github.com/repos/{repo.repo_full_name}/{safe_path}"
     response = github_request(method, url, access_token=access_token, **kwargs)
@@ -1215,7 +1213,7 @@ def _get_pr_info(github, repo_full_name: str, pr_number: int) -> dict:
     """
     from .github import github_request
 
-    access_token = github.integration.sensitive_config["access_token"]
+    access_token = github.get_access_token()
 
     response = github_request(
         "GET",
@@ -1250,7 +1248,7 @@ def _fetch_baseline_file(
 
     from .github import github_request
 
-    access_token = github.integration.sensitive_config["access_token"]
+    access_token = github.get_access_token()
 
     response = github_request(
         "GET",
@@ -1355,7 +1353,7 @@ def _post_commit_status(
         logger.debug("visual_review.status_check_skipped", run_id=str(run.id), reason="no_github_integration")
         return
 
-    access_token = github.integration.sensitive_config["access_token"]
+    access_token = github.get_access_token()
     target_url = f"{settings.SITE_URL}/project/{repo.team_id}/visual_review/runs/{run.id}"
 
     try:
